upd
upd

update docs

update doc

keep original code structure

fix version.py path

upd docs
