+++ .gitignore
@@ -12,7 +12,7 @@ __pycache__/
 # C extensions
 *.so
 
-build*
+build/
 
 *.ll
 .npm
